+++ StereoPlus2/ToolPool.hpp
@@ -9,7 +9,7 @@ class ToolPool {
 		tool->cross <<= Scene::cross();
 	}
 
-	static void Init(SinePenTool* tool) {
+	static void Init(CosinePenTool* tool) {
 		tool->cross <<= Scene::cross();
 	}